+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # CHANGELOG
 
 
+## 0.3.0
+
+* Turn the package into an ESM module (back again).
+  (Should work with Deno in the meanwhile.)
+* Update dependencies: no impact on shipped code.
+* Add functions: `when`, `withNewlineAppended`, and `commaSeparated`.
+* Rename `NestedString` to `Template`, leaving `NestedString` as an alias.
+  This is  not a breaking change, but I might deprecate and remove `NestedString` in the future.
+* Add {T|J}SDoc.
+
+
 ## 0.2.2
 
 * Fix internal `flatten` function for empty arrays.